Added WriteLogger backend
The WriteLogger backend is a write-only backend which prints all write
operations into a logfile.

+++ src/main/java/de/zib/gndms/vold/VolatileDirectoryImpl.java
@@ -110,7 +110,7 @@ public class VolatileDirectoryImpl implements VolatileDirectory
 
                 List< String > oldtimeslice = directory.lookup( 1, key );
 		
-		int newtimeslice = timeslice.getActualSlice();
+		long newtimeslice = timeslice.getActualSlice();
 
                 // insert new "slice/key |--> date" entry
                 {
